chore: address some feedback
